* 'devel' of https://github.com/arangodb/arangodb:
  Feature/remove manual zippery (#3036)
  Fixed issue with autoincrement data not being persisted properly in RocksDB (#3059). (#3067)
  Bug fix/issues 1708 (#3060)
  Fix issue #3037: Foxx, internal server error when I try to add a new service (#3056)
  Bug fix/v8 syslog (#3055)
  Bug fix/small issues 1608 (#3049)
  fixed issue #3044 (#3048)
  fixed issue #3039 (#3045)
  Fix foxx github url (#3042)
  we need to substitute the package name here too - else enterprise pac… (#3025)
  Fixing engine stats in arangosh (#3038)
  MSVC is pendantic (but right) (#3047)
  reduce log spam (#3051)
  Feature/build docker to build using stretch container (#3062)
  build docker to build using stretch container (#3061)
  fix and extend journal-related tests (#3043)

+++ arangod/Aql/Functions.cpp
@@ -1203,7 +1203,7 @@ AqlValue Functions::Lower(arangodb::aql::Query* query,
 
   AppendAsString(trx, adapter, value);
 
-  UnicodeString s(buffer->c_str(), buffer->length());
+  UnicodeString s(buffer->c_str(), static_cast<int32_t>(buffer->length()));
   s.toLower(NULL);
   s.toUTF8String(utf8);
 
@@ -1224,7 +1224,7 @@ AqlValue Functions::Upper(arangodb::aql::Query* query,
 
   AppendAsString(trx, adapter, value);
 
-  UnicodeString s(buffer->c_str(), buffer->length());
+  UnicodeString s(buffer->c_str(), static_cast<int32_t>(buffer->length()));
   s.toUpper(NULL);
   s.toUTF8String(utf8);
